+++ packages/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @asyra/core
 
+## 0.5.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @asyra/render@0.5.6
+
 ## 0.5.5
 
 ### Patch Changes